+++ examples/05_axios/src/DisplayRemoteData.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
+import axios from 'axios';
 
-import { useAsyncRun } from 'react-hooks-async';
-import { useAsyncTaskAxios } from 'react-hooks-async/src/use-async-task-axios';
+import { useAsyncRun, useAsyncTaskAxios } from 'react-hooks-async';
 
 const Err: React.SFC<{ error: Error }> = ({ error }) => (
   <div>Error:{error.name}{' '}{error.message}</div>
@@ -22,7 +22,7 @@ type Response = {
 
 const DisplayRemoteData: React.FC<{ id: string }> = ({ id }) => {
   const url = `https://jsonplaceholder.typicode.com/posts/${id}`;
-  const asyncTask = useAsyncTaskAxios<Response>({ url }, [url]);
+  const asyncTask = useAsyncTaskAxios<Response>(axios, { url }, [url]);
   useAsyncRun(asyncTask);
   const {
     pending,
