+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-hooks-async",
   "description": "An abortable async function library with React Hooks",
-  "version": "2.1.0",
+  "version": "3.0.0-alpha.0",
   "author": "Daishi Kato",
   "repository": {
     "type": "git",
