+++ src/index.d.ts
@@ -23,11 +23,67 @@ export type UseAsyncCombine = (
 
 // core async hooks
 
-export { useAsyncTask } from './use-async-task';
-export { useAsyncRun } from './use-async-run';
+export const useAsyncTask: UseAsyncTask;
+export const useAsyncRun: UseAsyncRun;
 
 // combining async hooks
 
-export { useAsyncCombineAll } from './use-async-combine-all';
-export { useAsyncCombineSeq } from './use-async-combine-seq';
-export { useAsyncCombineRace } from './use-async-combine-race';
+export const useAsyncCombineAll: UseAsyncCombine;
+export const useAsyncCombineSeq: UseAsyncCombine;
+export const useAsyncCombineRace: UseAsyncCombine;
+
+// timeout task
+
+export type UseAsyncTaskTimeout = <Result>(
+  func: () => Result,
+  delay: number,
+  deps: DependencyList,
+) => AsyncTask<Result>;
+
+export const useAsyncTaskTimeout: UseAsyncTaskTimeout;
+
+// delay task
+
+export type UseAsyncTaskDelay = (
+  milliSeconds: number,
+  deps: DependencyList,
+) => AsyncTask<true>;
+
+export const useAsyncTaskDelay: UseAsyncTaskDelay;
+
+// fetch task
+
+export type UseAsyncTaskFetch = <Result>(
+  input: string | Request,
+  init?: RequestInit,
+  bodyReader?: (b: Body) => Promise<Result>,
+) => AsyncTask<Result>;
+
+export const useAsyncTaskFetch: UseAsyncTaskFetch;
+export const useFetch: UseAsyncTaskFetch;
+
+// axios task
+
+export type UseAsyncTaskAxios = <
+  Result,
+  AxiosInstance = unknown,
+  AxiosRequestConfig = unknown,
+>(
+  axios: AxiosInstance,
+  config: AxiosRequestConfig,
+  deps: DependencyList,
+) => AsyncTask<Result>;
+
+export const useAsyncTaskAxios: UseAsyncTaskAxios;
+export const useAxios: UseAsyncTaskAxios;
+export default useAsyncTaskAxios;
+
+// wasm task
+
+export type UseAsyncTaskWasm = <Result>(
+  input: string | Request,
+  importObject?: object,
+) => AsyncTask<Result>;
+
+export const useAsyncTaskWasm: UseAsyncTaskWasm;
+export const useWasm: UseAsyncTaskWasm;
