+++ src/use-async-combine-all.js
@@ -22,5 +22,3 @@ export const useAsyncCombineAll = (...asyncTasks) => {
     result: asyncTasks.map(({ result }) => result),
   };
 };
-
-export default useAsyncCombineAll;
