--- src/use-async-run.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/named
-import { UseAsyncRun } from './index';
-
-export const useAsyncRun: UseAsyncRun;
-export default useAsyncRun;
