--- src/use-async-task-timeout.d.ts
@@ -1,11 +1,0 @@
-// eslint-disable-next-line import/named
-import { AsyncTask, DependencyList } from './index';
-
-export type UseAsyncTaskTimeout = <Result>(
-  func: () => Result,
-  delay: number,
-  deps: DependencyList,
-) => AsyncTask<Result>;
-
-export const useAsyncTaskTimeout: UseAsyncTaskTimeout;
-export default useAsyncTaskTimeout;
