--- src/use-async-task-wasm.d.ts
@@ -1,11 +1,0 @@
-// eslint-disable-next-line import/named
-import { AsyncTask } from './index';
-
-export type UseAsyncTaskWasm = <Result>(
-  input: string | Request,
-  importObject?: object,
-) => AsyncTask<Result>;
-
-export const useAsyncTaskWasm: UseAsyncTaskWasm;
-export const useWasm: UseAsyncTaskWasm;
-export default useAsyncTaskWasm;
