--- src/use-async-task.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/named
-import { UseAsyncTask } from './index';
-
-export const useAsyncTask: UseAsyncTask;
-export default useAsyncTask;
